+++ nodes/Autotask/operations/common/picklist-labels/index.ts
@@ -1,0 +1,3 @@
+import { picklistLabelOption, addPicklistLabelOption } from './description';
+
+export { picklistLabelOption, addPicklistLabelOption };